requirements.txt
faiss_cpu==1.8.0.post1
fastapi==0.115.6
pandas==2.2.3
pdfplumber==0.11.5
Pillow
pydantic==2.10.5
pytesseract==0.3.13
python_docx==1.1.2
sentence_transformers==3.0.1
transformers==4.44.0
PyMuPDF
pyngrok
python-multipart
accelerate
openai==1.59.7